requirements.txt
streamlit
llama-index
llama-index-vector-stores-chroma
chromadb
sentence-transformers
torch
transformers
nltk
groq
python-dotenv
llama-parse
pypdf